requirements.txt
aiohttp
aiofiles
asyncio
ffmpeg-python
Pillow
git+https://github.com/pyrogram/pyrogram@master
hachoir
lyricsgenius
py-tgcalls
wget
python-dotenv
youtube_search_python
youtube_search
speedtest-cli
motor
psutil
requests
youtube-search-python
rich
yt-dlp==2021.11.10.1 
pymongo[srv]
TgCrypto
